0.8.8
Added features
Inputs in duo mode
All the inputs that are available for the dedicated input mode have been added to the duo mode.

Flaps
The flaps have received a potentiometer focused function in the library and the connector.

//This function only takes the pin number of the potentiometer and handles the rest itself
connector.setPotFlaps(byte flapsPin);
This function will have a min and max range which are settable through the settings of the connector. This can be any number you’d like depending on the range of your potentiometer. (use Serial.println(analogRead(potPin)) to discover your own range). How the simulator handles this logic can differ per plane. If the flaps have set indents or stages, the potentiometer will trigger each stage, but you cannot put them in between indents (unless the plane facilitates this).

Throttles, mixture, props
The throttle, mixture and props have received some extra smoothness. The Arduino will only send differences that are bigger than 1 to ensure that the light jitter gets filtered out. It’s a simple solution for a simple problem but somehow it flew under my radar for far too long.

//functions like these should now reduce the load you are sending out significantly
connector.simpleInputHandling(throttlePin);
Bugfixes
Trim being saved as a byte. It now stores it as an int returning the negative values as well (-100 till 100)
Changing settings could crash the connector (forgot to implement a failsafe) this should be fixed (fingers crossed).
Mixture reverse range is now recognized (the setting groups weren’t opened / closed properly, It tried to search for the mixture in the engine section etc.)
Propeller reverse range is now recognized
Removed snedKhzComStandby1Plus and sendKhzNavStandby1Min (these were deprecated and were supposed to be removed a long time ago)
Still needs some investigating
For some reason the ADF is being a nosy brat. It sometimes messes with timings and data coming in. Haven’t pinpointed the issue yet.