Add presets feature to dashboard
Introduced a presets feature in the dashboard which allows saving and loading of configurations. This involved creating new files for handling presets, modifications to dashboard and settings handling classes along with necessary additions to CMakeLists.txt. The presets feature includes a new widget in the dashboard for showing preset rows, associated controllers and handlers, and new models for defining presets.