Add delay boot option to ComBundle
This commit adds a delay boot parameter to the ComBundle class constructor. This option is needed to allow for boards, such as Arduino Due, to reset after booting up. The ComPorts implementation has also been extended to automatically detect if a reset delay is necessary by checking the board type.