Refactor set handling and method naming
The static initialization of setList in SetHandler has been moved to the implementation file. This relocation aligns more with the standard practices. Method names in FormBuilder have been adjusted to increase clarity. 'retrieveSetting' and 'storeValue' are now known as 'getSettingValue' and 'storeSettingValue' respectively. Further changes in OutputMenu include reordering 'setCheckboxes' for better execution and the removal of superfluous whitespace.