Update OutputConverters and outputs.json with new type
OutputConverters has been updated to include a new case in the switch statement for KOHLSMANMODE. This change also supports converting and rounding off to two decimal places. Additionally, the type for "Kohlsman HG" in outputs.json is modified from 3 to 6. The updates support increased functionality and improve accuracy of computations.